requirements.txt
pip==24.3.1
bs4==0.0.2
requests==2.28.2
streamlit_card==1.0.2
streamlit_autorefresh==1.0.1